obkv-client:check tenant name after connect && insert null into auto-increment key (#63)
