+++ src/main/java/com/alipay/oceanbase/rpc/table/api/TableQuery.java
@@ -66,10 +66,12 @@ public interface TableQuery {
 
     /**
      * Row count offset, default: 0
-     *
+     * 
      * @param offset limit offset
      * @param limit limit count
      * @return this TableQuery
+     * 
+     * tips: can not use offset without valid limit value : limit < 0 && offset > 0
      */
     TableQuery limit(int offset, int limit);
